pyproject.toml
[project]
name = "dsc-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataset>=1.6.2",
    "datasets>=4.0.0",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "peft>=0.17.1",
    "py-vncorenlp>=0.1.4",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "transformers>=4.56.1",
]
